# docs/swagger.yaml
basePath: /
definitions:
  github_com_fluentfox_api_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_fluentfox_api_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorBody'
    type: object
...
  internal_chapter.ChapterDetailResponse:
    properties:
      concepts:
        items:
          $ref: '#/definitions/internal_chapter.ConceptResponse'
        type: array
      cultural_insights:
        items:
          $ref: '#/definitions/internal_chapter.CulturalInsightResponse'
        type: array
      description:
        type: string
      id:
        type: string
      jlpt_level:
        type: string
      order_index:
        type: integer
      slug:
        type: string
      title:
        type: string
    type: object
  internal_chapter.ChapterResponse:
    properties:
      description:
        type: string
      id:
        type: string
      jlpt_level:
        type: string
      order_index:
        type: integer
      slug:
        type: string
      title:
        type: string
    type: object
  internal_chapter.ConceptResponse:
    properties:
      explanation:
        type: string
      id:
        type: string
      order_index:
        type: integer
      title:
        type: string
    type: object
  internal_chapter.CulturalInsightResponse:
    properties:
      body:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  internal_quiz.AnswerResultResponse:
    properties:
      correct:
        type: boolean
      correct_answer:
        type: string
      explanation:
        type: string
      xp_earned:
        type: integer
    type: object
  internal_quiz.SessionResponse:
    properties:
      chapter_id:
        type: string
      id:
        type: string
      quiz_type:
        type: string
      total_cards:
        type: integer
    type: object
  internal_quiz.SessionSummaryResponse:
    properties:
      accuracy:
        type: number
      correct:
        type: integer
      incorrect:
        type: integer
      session_id:
        type: string
      xp_earned:
        type: integer
    type: object
  internal_quiz.StartSessionRequest:
    properties:
      card_count:
        maximum: 50
        minimum: 5
        type: integer
      chapter_id:
        type: string
      quiz_type:
        enum:
        - multiple_choice
        - fill_blank
        - listening
        type: string
    required:
    - quiz_type
    type: object
  internal_quiz.SubmitAnswerRequest:
    properties:
      content_id:
        type: string
      time_taken_ms:
        minimum: 0
        type: integer
      user_answer:
        type: string
    required:
    - content_id
    - time_taken_ms
    - user_answer
    type: object
  internal_shop.InventoryResponse:
    properties:
      id:
        type: string
      item_name:
        type: string
      item_type:
        type: string
      used_at:
        type: string
    type: object
  internal_shop.PurchaseRequest:
    properties:
      item_id:
        type: string
    required:
    - item_id
    type: object
  internal_shop.PurchaseResponse:
    properties:
      inventory_id:
        type: string
      message:
        type: string
    type: object
  internal_shop.ShopItemResponse:
    properties:
      description:
        type: string
      id:
        type: string
      item_type:
        type: string
      name:
        type: string
      xp_cost:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Japanese language learning API — chapters, SRS flashcard reviews, quizzes, XP levelling, streaks, and in-app shop.
    All successful responses are wrapped in {"data": ...}. Paginated responses also include {"meta": {...}}.
    All error responses follow {"error": {"code": "...", "message": "..."}}.
  title: FluentFox API
  version: "1.0"
paths:
...
  /chapters:
    get:
      description: Returns a paginated list of published chapters, optionally filtered
        by JLPT level.
      parameters:
      - description: JLPT level filter
        enum:
        - N5
        - N4
        - N3
        - N2
        - N1
        in: query
        name: level
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Results per page (default: 20)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_chapter.ChapterResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List chapters
      tags:
      - chapters
  /chapters/{slug}:
    get:
      description: Returns a single chapter with its grammar concepts and cultural
        insights.
      parameters:
      - description: Chapter slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chapter.ChapterDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chapter detail
      tags:
      - chapters
...
  /quiz/sessions:
    post:
      consumes:
      - application/json
      description: Creates a new quiz session of the given type. Optionally scoped
        to a specific chapter.
      parameters:
      - description: Quiz type, optional chapter, card count
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_quiz.StartSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_quiz.SessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a quiz session
      tags:
      - quiz
  /quiz/sessions/{id}/answers:
    post:
      consumes:
      - application/json
      description: Submits an answer for one item in an active quiz session. Returns
        correctness and XP earned.
      parameters:
      - description: Quiz session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Answer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_quiz.SubmitAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_quiz.AnswerResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a quiz answer
      tags:
      - quiz
  /quiz/sessions/{id}/finish:
    post:
      description: Marks the quiz session as completed and returns a summary with
        accuracy and total XP earned.
      parameters:
      - description: Quiz session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_quiz.SessionSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finish a quiz session
      tags:
      - quiz
  /shop:
    get:
      description: Returns all currently available items in the XP shop.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_shop.ShopItemResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List shop items
      tags:
      - shop
  /shop/inventory:
    get:
      description: Returns all items the authenticated user has purchased.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_shop.InventoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user inventory
      tags:
      - shop
  /shop/purchase:
    post:
      consumes:
      - application/json
      description: Spends XP to purchase an item. Returns inventory ID of the new
        item.
      parameters:
      - description: Item ID to purchase
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_shop.PurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_shop.PurchaseResponse'
        "400":
          description: Insufficient XP or item unavailable
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Purchase a shop item
      tags:
      - shop
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_access_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"